feat: move tables to their own components
