+++ src/app/v2/inventory/projects/[projectId]/page.tsx
@@ -3,14 +3,6 @@
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
 import { Card, CardContent } from "@/components/ui/card";
 import { Shield } from "lucide-react";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
 import VersionList from "@/components/projects/version-list";
 import {
   BOMWithAttributes,
@@ -18,19 +10,8 @@ import {
 } from "@buf/safedep_api.bufbuild_es/safedep/services/controltower/v1/project_pb";
 import { useEffect, useState } from "react";
 import { BOM_Status } from "@buf/safedep_api.bufbuild_es/safedep/messages/controltower/v1/bom_pb";
-
-function getBOMStatusName(status?: BOM_Status) {
-  switch (status) {
-    case BOM_Status.LATEST:
-      return "Latest";
-    case BOM_Status.HISTORICAL:
-      return "Historical";
-    case BOM_Status.UNSPECIFIED:
-      return "Unspecified";
-    default:
-      return "Unknown";
-  }
-}
+import ComponentsTable from "@/components/projects/components-table";
+import BOMSTable from "@/components/projects/boms-table";
 
 export default function ProjectDetails() {
   const versions = [
@@ -116,81 +97,19 @@ export default function ProjectDetails() {
       </div>
 
       {/* Tabs and Table */}
-      <Tabs defaultValue="boms" className="w-full">
+      <Tabs defaultValue="components" className="w-full">
         <TabsList className="mb-4 w-full grid grid-cols-2 lg:block lg:w-fit">
-          <TabsTrigger value="boms">BOMs</TabsTrigger>
           <TabsTrigger value="components">Components</TabsTrigger>
+          <TabsTrigger value="boms">BOMs</TabsTrigger>
         </TabsList>
 
         <TabsContent value="components">
-          <Table>
-            <TableHeader>
-              <TableRow>
-                <TableHead>Name</TableHead>
-                <TableHead>Version</TableHead>
-                <TableHead>Status</TableHead>
-              </TableRow>
-            </TableHeader>
-            <TableBody>
-              <TableRow>
-                <TableCell className="font-medium">dependency-1</TableCell>
-                <TableCell>v2.1.0</TableCell>
-                <TableCell>
-                  <span className="rounded-full bg-green-100 px-2 py-1 text-xs font-medium text-green-700">
-                    Secure
-                  </span>
-                </TableCell>
-              </TableRow>
-              <TableRow>
-                <TableCell className="font-medium">dependency-2</TableCell>
-                <TableCell>v1.8.3</TableCell>
-                <TableCell>
-                  <span className="rounded-full bg-green-100 px-2 py-1 text-xs font-medium text-green-700">
-                    Secure
-                  </span>
-                </TableCell>
-              </TableRow>
-              <TableRow>
-                <TableCell className="font-medium">dependency-3</TableCell>
-                <TableCell>v3.0.1</TableCell>
-                <TableCell>
-                  <span className="rounded-full bg-yellow-100 px-2 py-1 text-xs font-medium text-yellow-700">
-                    Review Required
-                  </span>
-                </TableCell>
-              </TableRow>
-            </TableBody>
-          </Table>
+          <ComponentsTable />
         </TabsContent>
 
         {/* BOM values */}
         <TabsContent value="boms">
-          <Table>
-            <TableHeader>
-              <TableRow>
-                <TableHead>BOM ID</TableHead>
-                <TableHead>Created At</TableHead>
-                <TableHead>Status</TableHead>
-                <TableHead>Last Updated</TableHead>
-              </TableRow>
-            </TableHeader>
-            <TableBody>
-              {boms.map(({ bom }) => (
-                <TableRow key={bom?.bomId}>
-                  <TableCell className="font-medium font-mono">
-                    {bom?.bomId}
-                  </TableCell>
-                  <TableCell>
-                    {bom?.createdAt?.toDate().toLocaleDateString()}
-                  </TableCell>
-                  <TableCell>{getBOMStatusName(bom?.status)}</TableCell>
-                  <TableCell>
-                    {bom?.updatedAt?.toDate().toLocaleDateString()}
-                  </TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
+          <BOMSTable boms={boms} />
         </TabsContent>
       </Tabs>
     </div>
